feat(frontend): make footer responsive
