+++ apps/frontends/omni-book/app/components/navigation/navbar.tsx
@@ -1,10 +1,10 @@
-import { Form, Link, NavLink, Outlet } from "react-router";
+import { Form, Link, NavLink, Outlet, useLocation } from "react-router";
 import { memo } from "react";
 import { Button } from "../ui/button";
 import { SearchIcon } from "../icons";
+import Footer from "../footer/footer";
 import type { Route } from "./+types/navbar";
 import { Profile } from "./profile";
-import Footer from "./footer";
 import type { paths } from "~/lib/api-types";
 import { env } from "~/lib/env";
 import { parseCookie } from "~/server/utils";
@@ -44,6 +44,7 @@ export const loader = async ({ request }: Route.LoaderArgs) => {
 };
 
 const Navbar = ({ loaderData }: Route.ComponentProps) => {
+  const location = useLocation();
   return (
     <AuthContext.Provider value={loaderData.user}>
       <nav className="sticky top-0 bg-background z-50">
@@ -120,7 +121,7 @@ const Navbar = ({ loaderData }: Route.ComponentProps) => {
         </div>
       </nav>
       <Outlet />
-      <Footer />
+      {!location.pathname.startsWith("/admin") && <Footer />}
     </AuthContext.Provider>
   );
 };
